final update of master from dev

+++ src/Controller/CategoryController.php
@@ -10,6 +10,7 @@
 namespace App\Controller;
 
 use App\Model\CategoryManager;
+use App\Model\AdminArticleManager;
 
 /**
  * Class ItemController
@@ -27,30 +28,21 @@ class CategoryController extends AbstractController
      * @throws \Twig\Error\RuntimeError
      * @throws \Twig\Error\SyntaxError
      */
-    // public function index()
-    // {
-    //     // $itemManager = new ItemManager();
-    //     // $items = $itemManager->selectAll();
 
-    //     return $this->twig->render('Category/category.html.twig');
-    // }
-
-
-    // *
-    //  * Display item informations specified by $id
-    //  *
-    //  * @param int $id
-    //  * @return string
-    //  * @throws \Twig\Error\LoaderError
-    //  * @throws \Twig\Error\RuntimeError
-    //  * @throws \Twig\Error\SyntaxError
      
-    public function show()
+    public function show($deleteNo = false)
     {
         $categoryManager = new CategoryManager();
-        $category = $categoryManager->selectAll();
+        $categories = $categoryManager->selectAll();
 
-        return $this->twig->render('Category/category.html.twig', ['categoryAll'=> $category]);
+        return $this->twig->render(
+            'Category/category_add.html.twig',
+            ['categoryAll'=> $categories,
+            'Btn' => 'Ajouter',
+            'method'=>'add',
+            'title_page' => ' Liste des catégories',
+            'deleteNo' =>$deleteNo]
+        );
     }
 
 
@@ -67,38 +59,38 @@ class CategoryController extends AbstractController
     {
         $categoryManager = new CategoryManager();
         $category = $categoryManager->selectOneById($id);
-
+        $categories = $categoryManager -> selectAll();
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $category['name'] = $_POST['name'];
             $categoryManager->update($category);
-            header('Location: /category/show');
+            header('Location: /Category/show');
         }
-
-        return $this->twig->render('Category/categoryEdit.html.twig', ['category' => $category]);
+        return $this->twig->render(
+            'Category/category_add.html.twig',
+            ['category' => $category,
+            'categoryAll'=> $categories,
+            'title_page' => 'Modifier catégorie',
+            'method'=>'edit/'.$category['id'],
+            'values' => $category,
+            'Btn' => 'Modifier']
+        );
     }
 
-
-//     *
+// *
 //      * Display item creation page
 //      *
 //      * @return string
 //      * @throws \Twig\Error\LoaderError
 //      * @throws \Twig\Error\RuntimeError
 //      * @throws \Twig\Error\SyntaxError
-     
     public function add()
     {
-
-
-      
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $categoryManager = new CategoryManager();
-            $addCat = ['addCat' => $_POST['addCat']];
-    
+            $addCat = ['name' => $_POST['name']];
             $idCat = $categoryManager -> insert($addCat);
             header('Location:/category/show');
         }
-
         return $this->twig->render('Category/category_add.html.twig');
     }
 
@@ -110,8 +102,14 @@ class CategoryController extends AbstractController
      */
     public function delete(int $id)
     {
-        $categoryManager = new CategoryManager();
-        $categoryManager->delete($id);
-        header('Location:/category/show');
+        $articleManager = new AdminArticleManager();
+        $nbArticles = $articleManager->checkArticlesOnCategory($id) ;
+        if ($nbArticles['nb'] == 0) {
+            $categoryManager = new CategoryManager();
+            $categoryManager->delete($id);
+            header('Location:/category/show');
+        } else {
+            header('Location:/category/show/deleteNo=true');
+        }
     }
 }
